+++ src/main/java/edu/ie3/util/quantities/QuantityUtilNew.java
@@ -5,7 +5,7 @@
 */
 package edu.ie3.util.quantities;
 
-import static edu.ie3.util.quantities.PowerSystemUnitsNew.DEGREE_GEOM;
+import static edu.ie3.util.quantities.PowerSystemUnits.DEGREE_GEOM;
 import static java.lang.StrictMath.abs;
 
 import javax.measure.Quantity;
@@ -74,7 +74,7 @@ public class QuantityUtilNew {
    * Compares two {@link Angle} {@link Quantity}s, if they are considerably equal. The comparison is
    * made on the absolute difference of both quantities' value. As of the repetitive nature of
    * angles, they have to be treated separately, e.g. -170° is semantically the same angle as 190°.
-   * To ensure this, all quantities are converted to {@link PowerSystemUnitsNew#DEGREE_GEOM}.
+   * To ensure this, all quantities are converted to {@link PowerSystemUnits#DEGREE_GEOM}.
    *
    * @param a First quantity to compare
    * @param b Second quantity to compare
